refactor: abstract tabs
