+++ pages/index.scss
@@ -1,109 +1,3 @@
-.tabs_container {
-    --ghost-padding: 20px;
-    display: block;
-    overflow: hidden;
-    position: sticky;
-    top: calc(var(--header-height, 0) - 2px);
-    background: rgba(255, 255, 255, 1);
-    z-index: 100;
-
-    .tabs {
-        --tab-vertical-padding: 20px;
-        display: flex;
-        width: 100%;
-        flex-flow: row;
-        animation: tabsFadeIn 1s cubic-bezier(0.28, 0.11, 0.32, 1) 1.2s;
-        animation-fill-mode: backwards;
-        max-width: 100vw;
-        overflow-y: hidden;
-        overflow-x: scroll;
-        padding-bottom: var(--ghost-padding);
-        margin-bottom: calc(var(--ghost-padding) * -1);
-        position: relative;
-
-        &:after {
-            content: "";
-            position: absolute;
-            right: 0;
-            bottom: var(--ghost-padding);
-            left: 0;
-            width: 100%;
-            height: 1px;
-            background-color: rgba(231, 232, 233, 0);
-            transition: background-color 0.2s ease;
-
-            @at-root [data-show-border="true"]#{&} {
-                background-color: rgba(231, 232, 233, 1);
-            }
-        }
-
-        >.current_tab_indicator {
-            position: absolute;
-            bottom: var(--ghost-padding);
-            left: 20px;
-            height: 2px;
-            background-color: rgb(0, 0, 0);
-            border-radius: 2px;
-            transform: translateX(var(--leading-offset, 0));
-            width: calc(var(--width, 0) - 40px);
-            z-index: 1;
-            padding: 0 16px;
-        }
-
-        .tab {
-            position: relative;
-
-            &:first-of-type {
-                margin-left: auto;
-            }
-
-            &:last-of-type {
-                margin-right: auto;
-            }
-
-            >input {
-                position: absolute;
-                top: 0;
-                left: 0;
-                right: 0;
-                bottom: 0;
-                width: 100%;
-                height: 100%;
-                padding: 0;
-                margin: 0;
-                -webkit-appearance: none;
-                -moz-appearance: none;
-                appearance: none;
-                outline: none;
-                border: none;
-                opacity: 0;
-                pointer-events: none;
-            }
-
-            >label {
-                display: block;
-                padding: var(--tab-vertical-padding) 20px;
-                cursor: pointer;
-                font-size: 15px;
-                font-weight: 460;
-                color: rgb(136, 139, 144);
-                transition: color 0.2s cubic-bezier(0.28, 0.11, 0.32, 1);
-                width: max-content;
-
-                @media (hover: hover) {
-                    &:hover {
-                        color: rgb(0, 0, 0);
-                    }
-                }
-            }
-
-            >input:checked+label {
-                color: rgb(0, 0, 0);
-            }
-        }
-    }
-}
-
 .feeds_container {
     overflow: hidden;
 
@@ -128,15 +22,5 @@
             margin: 0;
             padding: 0;
         }
-    }
-}
-
-@keyframes tabsFadeIn {
-    0% {
-        opacity: 0;
-    }
-
-    100% {
-        opacity: 1;
     }
 }
